+++ CHANGELOG.md
@@ -12,4 +12,5 @@
 * Add `c.draw.polylines` for optimized multiple-polyline rendering. It is possible to draw ~100k lines in 60 FPS using this function (instead of ~500 lines). Using this function, it was possible to implement:
   * `c.draw.ellipses` for optimized multiple ellipse drawing
   * `c.draw.rects` for optimized multiple rectangle drawing
+  * `c.draw.scatter` for scatter plots with fancy markers
 * Add experimental benchmarking setup